+++ CHANGELOG.md
@@ -18,6 +18,8 @@ speaking, considered to be the player's experience.
   on the Military Report.
 - Added a keybind for confirming in saving and loading menus (defaults to
   Enter), settable in config.txt with `bindkey = KEYEVENT_CONFIRM,Return`.
+- Added the KEYEVENT_CONFIRM keybind to mode selection screen for confirming the
+  selection using the Enter key (by default).
 
 ### Fixed